bigly_backend/requirements.txt
asgiref~=3.8.1
channels==4.1.0
daphne==4.1.2
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
fcm-django==2.2.1
ffmpeg-python==0.2.0
openai==1.35.10
Pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
social-auth-app-django==5.4.2
